+++ src/view/viewport.rs
@@ -1,36 +1,37 @@
-//! A view allows scrolling of a `Widget` without builtin
+//! A viewport allows scrolling of a `StatefulWidget` without builtin
 //! support for scrolling.
 //!
 //! View and Viewport are the same in functionality.
 //!
-//! The difference is that View works for [Widget]s and
+//! The difference is that View works for [Widget](ratatui::widgets::Widget)s and
 //! Viewport for [StatefulWidget]s.
 //!
 use crate::_private::NonExhaustive;
 use crate::event::ScrollOutcome;
 use crate::util::copy_buffer;
-use rat_event::{HandleEvent, MouseOnly, Outcome};
+use crate::view::ViewStyle;
+use rat_event::{flow, ConsumedEvent, HandleEvent, MouseOnly, Outcome};
 use rat_scrolled::{Scroll, ScrollArea, ScrollAreaState, ScrollState};
 use ratatui::buffer::Buffer;
 use ratatui::layout::{Rect, Size};
-use ratatui::prelude::StatefulWidget;
 use ratatui::style::Style;
-use ratatui::widgets::{Block, Widget};
 #[cfg(feature = "unstable-widget-ref")]
-use ratatui::widgets::{StatefulWidgetRef, WidgetRef};
+use ratatui::widgets::StatefulWidgetRef;
+use ratatui::widgets::{Block, StatefulWidget};
+use std::fmt::Debug;
 
-/// View has its own size, and can contain a stateless widget
+/// Viewport has its own size, and can contain a stateful widget
 /// that will be rendered to a view sized buffer.
 /// This buffer is then offset and written to the actual
 /// frame buffer.
 #[derive(Debug, Default, Clone)]
-pub struct View<'a, T> {
+pub struct Viewport<'a, T> {
     widget: Option<T>,
-    viewport: ViewImpl<'a>,
+    viewport: ViewportImpl<'a>,
 }
 
 #[derive(Debug, Default, Clone)]
-struct ViewImpl<'a> {
+struct ViewportImpl<'a> {
     block: Option<Block<'a>>,
     hscroll: Option<Scroll<'a>>,
     vscroll: Option<Scroll<'a>>,
@@ -40,7 +41,10 @@ struct ViewImpl<'a> {
 
 /// State of the viewport.
 #[derive(Debug, Clone)]
-pub struct ViewState {
+pub struct ViewportState<S> {
+    /// State of the viewed widget.
+    pub widget: S,
+
     /// Complete area of the viewport.
     /// __readonly__. renewed for each render.
     pub area: Rect,
@@ -62,7 +66,7 @@ pub struct ViewState {
     pub non_exhaustive: NonExhaustive,
 }
 
-impl<'a, T> View<'a, T> {
+impl<'a, T> Viewport<'a, T> {
     /// New viewport.
     pub fn new(inner: T) -> Self {
         Self {
@@ -98,6 +102,19 @@ impl<'a, T> View<'a, T> {
         self
     }
 
+    /// Combined style.
+    pub fn styles(mut self, styles: ViewStyle) -> Self {
+        self.viewport.style = styles.style;
+        if let Some(styles) = styles.scroll {
+            self.viewport.hscroll = self.viewport.hscroll.map(|v| v.styles(styles.clone()));
+            self.viewport.vscroll = self.viewport.vscroll.map(|v| v.styles(styles.clone()));
+        }
+        if styles.block.is_some() {
+            self.viewport.block = styles.block;
+        }
+        self
+    }
+
     /// Style for the empty space outside the rendered buffer.
     pub fn style(mut self, style: Style) -> Self {
         self.viewport.style = style;
@@ -106,17 +123,17 @@ impl<'a, T> View<'a, T> {
 }
 
 #[cfg(feature = "unstable-widget-ref")]
-impl<'a, T> StatefulWidgetRef for View<'a, T>
+impl<'a, T> StatefulWidgetRef for Viewport<'a, T>
 where
-    T: WidgetRef,
+    T: StatefulWidgetRef,
 {
-    type State = ViewState;
+    type State = ViewportState<T::State>;
 
     fn render_ref(&self, area: Rect, buf: &mut Buffer, state: &mut Self::State) {
         if let Some(w) = &self.widget {
             render_ref(
                 &self.viewport,
-                |area, buf| w.render_ref(area, buf),
+                |area, buf, state| w.render_ref(area, buf, state),
                 area,
                 buf,
                 state,
@@ -127,17 +144,17 @@ where
     }
 }
 
-impl<'a, T> StatefulWidget for View<'a, T>
+impl<'a, T> StatefulWidget for Viewport<'a, T>
 where
-    T: Widget,
+    T: StatefulWidget,
 {
-    type State = ViewState;
+    type State = ViewportState<T::State>;
 
     fn render(mut self, area: Rect, buf: &mut Buffer, state: &mut Self::State) {
         if let Some(w) = self.widget.take() {
             render_ref(
                 &self.viewport,
-                |area, buf| w.render(area, buf),
+                |area, buf, state| w.render(area, buf, state),
                 area,
                 buf,
                 state,
@@ -148,12 +165,12 @@ where
     }
 }
 
-fn render_ref(
-    viewport: &ViewImpl<'_>,
-    render_widget: impl FnOnce(Rect, &mut Buffer),
+fn render_ref<S>(
+    viewport: &ViewportImpl<'_>,
+    render_widget: impl FnOnce(Rect, &mut Buffer, &mut S),
     area: Rect,
     buf: &mut Buffer,
-    state: &mut ViewState,
+    state: &mut ViewportState<S>,
 ) {
     state.area = area;
 
@@ -188,7 +205,7 @@ fn render_ref(
     );
 
     let mut tmp = Buffer::empty(state.view);
-    render_widget(state.view, &mut tmp);
+    render_widget(state.view, &mut tmp, &mut state.widget);
 
     copy_buffer(
         state.view,
@@ -201,9 +218,10 @@ fn render_ref(
     );
 }
 
-impl Default for ViewState {
+impl<S: Default> Default for ViewportState<S> {
     fn default() -> Self {
         Self {
+            widget: S::default(),
             area: Default::default(),
             inner: Default::default(),
             view: Default::default(),
@@ -214,11 +232,38 @@ impl Default for ViewState {
     }
 }
 
-impl ViewState {
-    pub fn new() -> Self {
-        Self::default()
+impl<S> ViewportState<S>
+where
+    S: Default,
+{
+    pub fn new(widget_state: S) -> Self {
+        Self {
+            widget: widget_state,
+            ..Self::default()
+        }
     }
+}
 
+impl<S> ViewportState<S> {
+    /// Relocate mouse-events for use inside the viewport.
+    pub fn relocate_crossterm(&self, event: &crossterm::event::Event) -> crossterm::event::Event {
+        match event {
+            crossterm::event::Event::FocusGained => event.clone(),
+            crossterm::event::Event::FocusLost => event.clone(),
+            crossterm::event::Event::Key(_) => event.clone(),
+            crossterm::event::Event::Mouse(m) => {
+                let mut m = *m;
+                m.column += self.hscroll.offset() as u16;
+                m.row += self.vscroll.offset() as u16;
+                crossterm::event::Event::Mouse(m)
+            }
+            crossterm::event::Event::Paste(_) => event.clone(),
+            crossterm::event::Event::Resize(_, _) => event.clone(),
+        }
+    }
+}
+
+impl<S> ViewportState<S> {
     pub fn vertical_offset(&self) -> usize {
         self.vscroll.offset()
     }
@@ -272,8 +317,16 @@ impl ViewState {
     }
 }
 
-impl HandleEvent<crossterm::event::Event, MouseOnly, Outcome> for ViewState {
-    fn handle(&mut self, event: &crossterm::event::Event, _qualifier: MouseOnly) -> Outcome {
+impl<R, Q, S> HandleEvent<crossterm::event::Event, Q, R> for ViewportState<S>
+where
+    S: HandleEvent<crossterm::event::Event, Q, R>,
+    R: From<Outcome> + ConsumedEvent + Debug,
+{
+    fn handle(&mut self, event: &crossterm::event::Event, qualifier: Q) -> R {
+        flow!(self
+            .widget
+            .handle(&self.relocate_crossterm(event), qualifier));
+
         let mut sas = ScrollAreaState::new()
             .area(self.inner)
             .h_scroll(&mut self.hscroll)
@@ -290,9 +343,9 @@ impl HandleEvent<crossterm::event::Event, MouseOnly, Outcome> for ViewState {
             ScrollOutcome::Changed => true,
         };
         if r {
-            return Outcome::Changed;
+            return Outcome::Changed.into();
         }
 
-        Outcome::Continue
+        Outcome::Continue.into()
     }
 }